Moved MessageManagers from entity to use case
