+++ src/main/java/use_case_message/MessageInteractor.java
@@ -1,7 +1,5 @@
 package use_case_message;
 
-import entity.MessageManagers;
-
 import java.sql.Timestamp;
 import java.text.SimpleDateFormat;
